refactor(helper): split mino-sandbox-helper into multi-file binary crate
Migrate from a 1850-line monolith to a structured 8-module crate under
src/bin/mino-sandbox-helper/ for auditability without changing behaviour.

Module layout:
- main.rs         — dispatch, load_request(), respond_healthy(), print helpers
- acl.rs          — set_acl(), cleanup_acls(), remove_acl()
- lifecycle.rs    — SpawnParams, SpawnGuard, SpawnStage, SPAWN_STAGES,
                    handle_spawn(), prepare_spawn(), get_user_ids(),
                    chown_recursive(), drop_privileges(), child/parent process
- resource_limits.rs — apply_resource_limits(), set_rlimit(), RlimitResource
- signal.rs       — CHILD_PID static, setup_signal_forwarding(), forward_signal()
- dotfiles.rs     — copy_dotfiles(), recreate_symlink_entry()
- exec.rs         — exec_command(), handle_exec(), ExecArgs, parse_exec_args()
- cleanup.rs      — handle_cleanup(), dispatch_cleanup(), CleanupArgs,
                    parse_cleanup_args()

All tests move with their functions into per-module #[cfg(test)] blocks.
Binary name remains mino-sandbox-helper — sudoers path is unchanged.
Cargo.toml path updated from .rs file to directory/main.rs.
